:sparkles: Add NonZeroValidator for the non-zero integer range
Add NonZeroValidator and validate_non_zero, rejecting a value of 0 -- the one
integer range Django's MinValueValidator/MaxValueValidator cannot express as a
single validator. It pins its deconstruct path to django_boost.validators so
migrations stay stable regardless of the internal module layout.

Split the flat validators module into a package (json, color, collection, uuid,
integer) that re-exports the public API unchanged, and track each submodule in
the mypy typed-module registry so the split keeps the previous type coverage.

Refs: https://github.com/ChanTsune/django-boost/issues/164
