+++ mypy.ini
@@ -97,6 +97,12 @@ disallow_untyped_defs = True
 disallow_incomplete_defs = True
 warn_return_any = True
 
+[mypy-django_boost.validators.*]
+ignore_errors = False
+disallow_untyped_defs = True
+disallow_incomplete_defs = True
+warn_return_any = True
+
 [mypy-django_boost.test.testcase]
 ignore_errors = False
 disallow_untyped_defs = True
